chore: move mapping common logic to shared function (#2492)
